Fix go.mod module declaration.
I'm not sure how this hasn't come up before, but if you try to use go get to grab the project, you get the following error:

go get: github.com/casbin/casdoor@none updating to
        github.com/casbin/casdoor@v0.0.0-20210720132906-66f36d780ad8: parsing go.mod:
        module declares its path as: github.com/casdoor/casdoor
                but was required as: github.com/casbin/casdoor

I'm not a big go user, but I believe this is how it's supposed to be.

Solves the issue #25. Seems I wasn't the first to run into the issue, but it wasn't looked to far into at the time.